backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "industrial-request-intelligence-platform-backend"
version = "0.1.0"
description = "Backend base for the Industrial Request Intelligence Platform"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.14.0,<2.0.0",
    "pypdf>=5.3.0,<6.0.0",
    "PyMuPDF>=1.27.0,<2.0.0",
    "rapidocr-onnxruntime==1.2.3",
    "fastapi>=0.115.0,<1.0.0",
    "email-validator>=2.2.0,<3.0.0",
    "pydantic-settings>=2.6.0,<3.0.0",
    "python-multipart>=0.0.9,<1.0.0",
    "redis>=5.2.0,<6.0.0",
    "sqlalchemy>=2.0.36,<3.0.0",
    "asyncpg>=0.30.0,<1.0.0",
    "psycopg[binary]>=3.2.0,<4.0.0",
    "uvicorn[standard]>=0.32.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
    "aiosqlite>=0.20.0,<1.0.0",
    "httpx>=0.28.0,<1.0.0",
    "Pillow>=12.1.0,<13.0.0",
    "pytest>=8.3.0,<9.0.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["app*"]
